requirements.txt
Django==1.8.4
djangorestframework==3.2.3
Markdown==2.6.2
Pillow==2.9.0